Complete lesson 02-flex-header
The lesson was done twice due to a misunderstanding. Too many
flexboxes were initially used in an attempt to center the header at
all times, as the desired-outcome pngs depicted. A fresh attempt was
done to gain a clear understanding of the concepts.

Redesignated the original attempt as "copy-02-flex-header", and did
not reference solution.css during the second attempt to promote more
efficient use of flex.